Issue #2 & Issue #7
Issue #2 - I'm creating a seperate line chart for each arm.  Issue #7 -
I created a data utility class. So far it has methods to get bar data
(not yet implemented) and line data. This class was created because the
file reader class should only be responsible for reading from a file and
not formatting the data.