Issue #16 - Game client.
Added the game client view and integrated it with the rest of the app.
Graph rendering needs to be fixed, instead of creating a new graph each time,
we should just be updating the current one. Also, when a new arm is added,
it is not displayed without refreshing. Same goes with when an arm is deleted.